Reduce measurement noise
When using the make plot command to measure the execution time, the
do_measurement script will be invoked to setup the following settings in
order to minimize the measurement noise.
 - disable address space layout randomization
 - set cpu governor to performance
 - disable turbo mode on intel cpu

All changes are temporary. The do_measurement script will change back to
their original settings after the measurement is done.