Implement statistical calculation on measurement
Originally, the client_plot only measures the runtime of fibdrv once. As
a result, the difference between each measurement can be significant and
leads to the unrepresentative of the result.
In order to overcome this issue, the client_statistic measures the
runtime of fibdrv several times and removes outliers based on the 95%
confidence interval.
The 'make plot' command will now support ploting 2 different results
based on the raw data from client_plot and the statistic data from
client_statistic respectively.